chore: move test files to __tests__ directory
