+++ src/scripts/__tests__/petstore-expanded.yaml
